Add player and AI move interface
Added support for populating the game board with the player and AI moves
When the user makes a move a POST call is sent to the server so the AI can
see it and make it's move in response. The AI isn't programmed yet, so
always selects the same tile.

Also cleaned up look of the client and added a few comments.